+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added support to ```javascript```library.
 - Added support to ```caching``` library.
 - Added support to ```file upload``` library.
+- Added support to ```calendar``` library.
 
 
 ## [0.0.4]